[Refactor] Refactor login and signup components and reuse logic with custom hook
