+++ package.json
@@ -42,8 +42,7 @@
     "styled-components": "^5.3.1",
     "supercluster": "^7.1.4",
     "use-supercluster": "^0.3.0",
-    "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "uuid": "^8.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
